auth-service/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
supabase
python-multipart
asyncpg
bcrypt
httpx
python-dotenv
email-validator
google-cloud-pubsub